Bumping code coverage thresholds back up

+++ generators/src/main/java/com/pholser/junit/quickcheck/generator/java/lang/CharacterGenerator.java
@@ -67,7 +67,7 @@ public class CharacterGenerator extends Generator<Character> {
 
     @Override public List<Character> doShrink(SourceOfRandomness random, Character larger) {
         return new CodePointShrink(cp -> cp >= min && cp <= max)
-            .shrink(random, larger)
+            .shrink(random, (int) larger)
             .stream()
             .map((Integer cp) -> (char) cp.intValue())
             .collect(Collectors.toList());
